Build large maps more efficiently
Literal maps in Go are initialized at runtime, and the code for every
entry of the map is individually generated. This code can be very
large and very branchy, especially if the map items require lots of
memory allocations.

While this is mostly not a problem, it does bloat executable size
compared to the same literal data in array form. And more importantly,
the extremely branchy nature of the code causes the WASM code to be
truly insane, with thousands and thousands of nested blocks. This can
cause problems; we found that TurboFan, v8's optimizing compiler, would
take several minutes and over 6GB of memory processing these large
functions. This exceeded a memory limit in Chrome and killed the tab.

Specifying all the data as an array of structs allows for the data to be
part of the data section of the executable, rather than generated by
code. We can then loop over that data on init to produce a map, much
like the original codegen would have done, but this time the generated
code is very small and simple, avoiding any optimization issues.